backend: necessary corrections and axios config
