+++ backend/.env.example
@@ -1,8 +1,4 @@
-#project ID
 PROJECT_ID
-#your dataset
 DATASET
-#token
 AUTH_TOKEN
-#Application port
 PORT
