+++ backend/sanity.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'sanity';
 import { deskTool } from 'sanity/desk';
 import { visionTool } from '@sanity/vision';
-import schemaTypes from './schemas';
+import { schemaTypes } from './schemas';
 
 export default defineConfig({
   name: 'default',
