+++ backend/schemas/index.ts
@@ -1,5 +1,3 @@
 import { introduction } from './introduction';
 
-const schemaTypes = [introduction];
-
-export default schemaTypes;
+export const schemaTypes = [introduction];
